+++ lib/hive_plus_secure.dart
@@ -1,6 +1,8 @@
 /// Hive Plus is a lightweight and blazing fast key-value store made for\n/// Flutter and Dart. It is strongly encrypted using AES-256.
 library hive_plus_secure;
 
+import 'dart:async';
+
 import 'package:hive_plus_secure/src/impl/frame.dart';
 import 'package:isar_plus/isar_plus.dart';